[AD-543] build error fix: change JDBC type from int8_t to int16_t
* some of the values of JDBC Type (e.g.,2009, 2012) would overflow if the value is int8_t.
* return type for function SqlTypeToBinary (which returns JDBC type) is changed to int16_t as well

+++ src/odbc/include/ignite/odbc/type_traits.h
@@ -191,7 +191,8 @@ namespace ignite
              * @param sqlType SQL type.
              * @return Binary type.
              */
-            int8_t SqlTypeToBinary(int16_t sqlType);
+            //int8_t SqlTypeToBinary(int16_t sqlType); // todo remove -al-
+            int16_t SqlTypeToBinary(int16_t sqlType);
 
             /**
              * Convert ODBC type to driver type alias.
